new pairs.utils.ts file to correct imports into bonds subgraph
