+++ pairs/src/Pair.ts
@@ -20,7 +20,8 @@ import { Pair, Token, Swap } from '../generated/schema'
 import { Swap as SwapEvent, Pair as PairContract } from '../generated/KLIMA_USDC/Pair'
 import { ERC20 as ERC20Contract } from '../generated/KLIMA_USDC/ERC20'
 import { Address } from '@graphprotocol/graph-ts'
-import { BigDecimalZero, BigIntZero, calculateCCO2AdjustedPrice } from './utils'
+import { BigDecimalZero, BigIntZero } from './utils'
+import { calculateCCO2AdjustedPrice } from './pair.utils'
 import { hourTimestamp } from '../../lib/utils/Dates'
 import { PriceUtil } from '../../lib/utils/Price'
